requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.124.2
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.20
regex==2025.11.3
requests==2.32.5
semantic-version==2.10.0
setuptools==80.9.0
setuptools-rust==1.12.0
starlette==0.50.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uuid==1.30
uvicorn==0.38.0